+++ web/packages/shared/utils/errorType.ts
@@ -14,9 +14,19 @@
  * limitations under the License.
  */
 
-import { assertUnreachable } from 'shared/utils/error';
+import { privateKeyEnablingPolicies } from 'shared/services';
 
-/**
- * @deprecated Import assertUnreachable from `shared/utils/error` instead.
- */
-export { assertUnreachable };
+export function isPrivateKeyRequiredError(err: Error) {
+  return privateKeyEnablingPolicies.some(p => err.message.includes(p));
+}
+
+// getErrMessage first checks if the error is of type Error
+// before attempting to access the error message field.
+// Used with try catch blocks, where the error caught
+// may not necessary be of type Error.
+export function getErrMessage(err: unknown) {
+  let message = 'something went wrong';
+  if (err instanceof Error) message = err.message;
+
+  return message;
+}
