+++ web/packages/teleport/src/Discover/Server/EnrollEc2Instance/EnrollEc2Instance.tsx
@@ -19,7 +19,7 @@ import { Box, Text } from 'design';
 import { FetchStatus } from 'design/DataTable/types';
 import useAttempt from 'shared/hooks/useAttemptNext';
 
-import { getErrMessage } from 'shared/utils/error';
+import { getErrMessage } from '@gravitational/shared/utils/errorType';
 
 import cfg from 'teleport/config';
 import { NodeMeta, useDiscover } from 'teleport/Discover/useDiscover';
